% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevators.R
\docType{data}
\name{elevators}
\alias{elevators}
\title{Elevators in NYC}
\format{
A data frame with 40140 rows and 27
variables:
}
\source{
\url{https://github.com/datanews/elevators}
}
\usage{
elevators
}
\description{
A data set containing information of a subset of the elevators in NYC.
}
\keyword{datasets}
