% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevators.R
\docType{data}
\name{elevators_raw}
\alias{elevators_raw}
\title{Elevators in NYC}
\format{
A data frame with 76088 rows and
29 variables:
}
\source{
\url{https://github.com/datanews/elevators}
}
\usage{
elevators_raw
}
\description{
A raw data set containing information of the elevators in NYC.
}
\keyword{datasets}
