% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_elevators.R
\docType{data}
\name{md_elevators}
\alias{md_elevators}
\title{Elevators in Maryland}
\format{
A data frame with 25204 rows and
15 variables:

\describe{
\item{location_name}{}
\item{location_type}{}
\item{county}{}
\item{location_address}{}
\item{registration_number}{}
\item{status}{}
\item{certificate_expiration_date}{}
\item{equipment_type}{}
\item{equipment_capacity}{}
\item{equipment_speed}{}
\item{equipment_travel}{}
\item{owner_contact}{}
\item{owner_contact_2}{}
\item{inspection_type}{}
\item{last_inspection_date}{}
}
}
\source{
\url{https://www.dllr.state.md.us/labor/safety/elevcounties.shtml}
}
\usage{
md_elevators
}
\description{
A data set containing information of a subset of the elevators in Maryland.
Last updated July 28th 2022.
}
\keyword{datasets}
